+++ packages/sdk/src/models/did-management/key-pair-did.ts
@@ -1,0 +1,8 @@
+import { ec, eddsa } from 'elliptic';
+
+export interface IKeyPairDID {
+  keyPair: eddsa.KeyPair | ec.KeyPair;
+  sign(message: Uint8Array): Uint8Array;
+  getPublicKeyBase58(): string;
+  getPublicKeyHex(): string;
+}
